requirements.txt
aniso8601==10.0.0
attrs==24.3.0
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
coloredlogs==15.0.1
filelock==3.17.0
flasgger==0.9.7.1
Flask==3.1.0
Flask-Cors==5.0.0
flask-restx==1.3.0
flatbuffers==25.1.21
fsspec==2024.12.0
humanfriendly==10.0
idna==3.10
imageio==2.37.0
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lazy_loader==0.4
llvmlite==0.44.0
MarkupSafe==3.0.2
mistune==3.1.0
mpmath==1.3.0
networkx==3.4.2
numba==0.61.0
numpy==2.1.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
onnxruntime==1.20.1
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==24.2
pillow==11.1.0
platformdirs==4.3.6
pooch==1.8.2
protobuf==5.29.3
PyMatting==1.1.13
pytz==2024.2
PyYAML==6.0.2
referencing==0.36.1
rembg==2.0.61
requests==2.32.3
rpds-py==0.22.3
scikit-image==0.25.0
scipy==1.15.1
six==1.17.0
sympy==1.13.1
tifffile==2025.1.10
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
triton==3.1.0
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
